added working styles.js
